refactor: split sleeper_tools — extract strategic-planning cluster (+ fix latent error-path bug)
Move the ~670-line strategic-planning tools (get_strategic_matchup_preview, get_season_bye_week_coordination, get_trade_deadline_analysis, get_playoff_preparation_plan) out of sleeper_tools.py into nfl_mcp/sleeper_strategy.py. Unlike enrichment (a leaf), these are top-level consumers that call core tools (get_league, get_matchups), so sleeper_tools re-exports them at the END of the module (after those names are defined) to keep the import acyclic; sleeper_strategy imports get_league/get_matchups from core. sleeper_tools.py: 2026 -> 1372 lines (3022 -> 1372 across both splits). Repointed the strategy white-box tests to patch sleeper_strategy (where the moved functions resolve their collaborators). FIX: the split's tests surfaced a latent crash — all 4 strategic tools passed their fallback data dict POSITIONALLY to create_error_response(), colliding with the error_type= kwarg (TypeError) on any upstream failure; now pass data= + added a regression test. ruff clean; full suite 902 passed, 2 skipped.
